requirements.txt
#docxtpl==0.17.0
Flask==3.0.3
Flask-RESTful==0.3.10
flask-cors
Flask-Caching
langchain
langchain_community
langchain_core
langchain_openai
#langchain-text-splitters==0.2.2
langchain-text-splitters
langsmith
langgraph
openai
llama_index==0.12.33
llama-index-core
llama_parse
tavily-python
pymupdf4llm
pymupdf==1.25.2
python-docx>=0.8.11
reportlab>=3.6.0
pinecone-client[grpc]==6.0.0
llama-index-vector-stores-pinecone
pymongo
lz4==4.3.2
tiktoken
gunicorn
werkzeug
python-magic
firebase-admin
flask-jwt-extended
fuzzywuzzy[speedup]
nltk
celery
redis